backend/requirements.txt
daphne==4.0.0
Django==4.2.4
django-cors-headers==4.2.0
django-nextjs==2.3.1
djangorestframework==3.14.0
python-dotenv==1.0.0
